rename _neighbors_at_link module; remove unused _link.py
