+++ cython-files.txt
@@ -10,7 +10,7 @@ src/landlab/components/flow_accum/cfuncs.pyx
 src/landlab/components/flow_director/cfuncs.pyx
 src/landlab/components/flow_router/ext/single_flow/priority_routing/_test_breach_c.pyx
 src/landlab/components/flow_router/ext/single_flow/priority_routing/breach.pyx
-src/landlab/components/overland_flow/_neighbors_at_link.pyx
+src/landlab/components/overland_flow/_calc.pyx
 src/landlab/components/priority_flood_flow_router/cfuncs.pyx
 src/landlab/components/space/ext/calc_qs.pyx
 src/landlab/components/space/ext/calc_sequential_ero_depo.pyx
