+++ tests/components/overland_flow/test_parallel_links.py
@@ -3,7 +3,7 @@ import pytest
 from numpy.testing import assert_array_equal
 
 from landlab import RasterModelGrid
-from landlab.components.overland_flow._neighbors_at_link import sum_parallel_links
+from landlab.components.overland_flow._calc import sum_parallel_links
 from landlab.graph.structured_quad.ext.at_link import fill_parallel_links_at_link
